pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vox"
version = "0.1.0"
description = "AI voice input tool for Windows - local STT + LLM text formatting"
requires-python = ">=3.11"
dependencies = [
    "pynput>=1.7",
    "sounddevice>=0.4",
    "numpy>=1.24",
    "faster-whisper>=1.0",
    "openai>=1.0",
    "pyperclip>=1.8",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
vox = "vox.__main__:main"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
